# openapi.yaml
openapi: 3.1.0
info:
  title: saas-scaffolding
  version: 1.0.0
  description: Multi-tenant SaaS scaffolding API with authentication, organization management, microservice tokens, and webhook integrations.
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie set after successful sign-in.
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Internal server error
...
    CreateOrganizationRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          example: Acme Corp
        slug:
          type: string
          description: URL-friendly identifier.
          example: acme-corp
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
    OrganizationWithRole:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
        role:
          type: string
          enum:
            - owner
            - admin
            - member
...
    AvatarUploadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Avatar uploaded successfully
        avatarUrl:
          type: string
          format: uri
    AvatarDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Avatar removed successfully
    StripeWebhookResponse:
      type: object
      properties:
        received:
          type: boolean
          example: true
  responses:
    BadRequest:
      description: Bad request - validation error or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or session invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations for the current user
      description: |
        Returns organizations the authenticated user is an active member of, with role information.
        Rate limit tier: apiRateLimit.
      tags:
        - organizations
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Array of organizations with roles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationWithRole"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      operationId: createOrganization
      summary: Create a new organization
      description: |
        Creates a new organization and makes the current user the owner. The slug must be unique and pass tenant slug validation.
        Rate limit tier: apiRateLimit.
      tags:
        - organizations
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Organization slug already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Organizations - Update (rename)
  # ──────────────────────────────────────────────
...
  /api/users/avatar:
    post:
      operationId: uploadAvatar
      summary: Upload user avatar
      description: |
        Uploads a new avatar image for the authenticated user. Replaces the previous avatar if one exists.
        Rate limit tier: uploadRateLimit.
      tags:
        - users
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (PNG, JPG, etc.).
      responses:
        "200":
          description: Avatar uploaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteAvatar
      summary: Remove user avatar
      description: |
        Deletes the authenticated user's avatar image and clears the avatar URL.
        Rate limit tier: uploadRateLimit.
      tags:
        - users
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Avatar removed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarDeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Webhooks - Stripe
  # ──────────────────────────────────────────────
  /api/webhooks/stripe:
    post:
      operationId: stripeWebhook
      summary: Stripe webhook receiver
      description: |
        Receives and verifies Stripe webhook events using the `stripe-signature` header.
        No session authentication; secured by Stripe signature verification.
        No rate limiting applied (Stripe controls delivery rate).
      tags:
        - webhooks
      security: []
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
          description: Stripe webhook signature for payload verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Stripe event payload (consumed as text for signature verification).
      responses:
        "200":
          description: Webhook received and processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeWebhookResponse"
        "400":
          description: Missing or invalid signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"